实现回复列表功能

+++ src/main/java/com/hh/community/community/Model/Question.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Question {
 
-    private Integer id;
+    private Long id;
 
     private String title;
 
@@ -15,7 +15,7 @@ public class Question {
 
     private Long gmtModified;
 
-    private Integer creator;
+    private Long creator;
 
     private Integer commentCount;
